Added led detection and switched to 4 space indentation
