+++ source/resource/bit_mask.py
@@ -1,0 +1,42 @@
+import cv2
+import numpy as np
+
+
+def gray_to_thresh(image, thresh_val):
+    gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+    blurred = cv2.GaussianBlur(gray, (5, 5), 0)
+    thresh = cv2.threshold(blurred, thresh_val, 255, cv2.THRESH_BINARY)[1]
+    return thresh
+
+
+def saliency_to_thresh(image, threshold):
+    saliency = cv2.saliency.StaticSaliencyFineGrained_create()
+    _, saliency_map = saliency.computeSaliency(image)
+    saliency_map = (saliency_map * 255).astype("uint8")
+
+    thresh = cv2.threshold(saliency_map.astype("uint8"), threshold, 255,
+                           cv2.THRESH_BINARY)[1]
+    return thresh
+
+
+def find_led_from_image(frame):
+    # layer 1: Only high hue pixels using hsv color format
+    hsv = cv2.cvtColor(frame, cv2.COLOR_BGR2HSV)
+    layer1 = cv2.inRange(hsv, (179 * 0, 0, 250), (179 * 0.5, 60, 255))
+    kernel = np.ones((2, 2), np.uint8)
+    layer1 = cv2.dilate(layer1, kernel, iterations=7)
+
+    # layer 2: Everything near bright red
+    layer2 = cv2.inRange(frame, (0, 0, 250), (150, 150, 255))
+    kernel = np.ones((4, 4), np.uint8)
+    layer2 = cv2.dilate(layer2, kernel, iterations=10)
+
+    # layer 3: only bright pixels
+    layer3 = cv2.inRange(frame, (200, 200, 253), (255, 255, 255))
+    kernel = np.ones((3, 3), np.uint8)
+    layer3 = cv2.dilate(layer3, kernel, iterations=2)
+
+    # final image: only the overlap of all 3 layers
+    result = cv2.bitwise_and(layer1, layer2)
+    result = cv2.bitwise_and(result, layer3)
+    return result
