+++ source/slope_thresh_led_image.py
@@ -6,25 +6,27 @@ import cv2
 # construct the argument parser and parse the arguments
 ap = argparse.ArgumentParser()
 ap.add_argument("-i", "--image", required=True,
-	help="path to the input image")
-ap.add_argument("--save", required=True, 
-	help="Path to saved image directory")
+                help="path to the input image")
+ap.add_argument("--save", required=True,
+                help="Path to saved image directory")
 args = vars(ap.parse_args())
 
+
 # function for saving image to disk in "save" directory
-def save_to_disk(img, img_name): 
-	idx = img_name.find('image-')
-	if args['image'][len(args['image'])-1] == '/':
-		path = args["save"] + img_name[idx:]
-	else: 
-		path = args["save"] + '/' + img_name[idx:]
-	cv2.imwrite(path, img)
+def save_to_disk(img, img_name):
+    idx = img_name.find('image-')
+    if args['image'][len(args['image']) - 1] == '/':
+        path = args["save"] + img_name[idx:]
+    else:
+        path = args["save"] + '/' + img_name[idx:]
+    cv2.imwrite(path, img)
+
 
 # 1. preprocess
-width=600
+width = 600
 image = cv2.imread(args["image"])
 image = imutils.resize(image, width)
-cv2.imshow('resized',image)
+cv2.imshow('resized', image)
 cv2.waitKey(0)
 
 # 2. Mask
@@ -40,12 +42,12 @@ target_leds = match_leds.slope_thresh(quad_centers, width)
 
 # 5. Compute Center
 coords = find_center.find_target_center(target_leds)
-print('Center Coordinates: '+str(coords))
+print('Center Coordinates: ' + str(coords))
 
 # show coordinate on image
 cv2.circle(image, (coords[0], coords[1]), 3, (0, 255, 0), -1)
-cv2.putText(image, "center", (coords[0]-25, coords[1]+20), 
-	cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0), 2)
+cv2.putText(image, "center", (coords[0] - 25, coords[1] + 20),
+            cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0), 2)
 cv2.imshow('center', image)
 cv2.waitKey(0)
 save_to_disk(image, args['image'])
